refactor: move Transaction to lance-table
