+++ rust/lance-table/src/format.rs
@@ -8,6 +8,7 @@ use uuid::Uuid;
 mod fragment;
 mod index;
 mod manifest;
+pub mod transaction;
 
 pub use fragment::*;
 pub use index::Index;
